+++ cypress/integration/MaskedInput/index.spec.js
@@ -10,10 +10,7 @@ describe('MaskedInput', () => {
   describe('Interaction', () => {
     describe('Input', () => {
       it('should clear one char per backspace press', () => {
-        cy.get(`.${theme.wrapper}`)
-          .eq(0)
-          .get(`.${theme.input}`)
-          .eq(0)
+        cy.name('MINotControlledPhone')
           .focusMasked()
           .clear()
           .type('9818862798')
@@ -37,9 +34,27 @@ describe('MaskedInput', () => {
           .should('have.value', '+7 (___)-___-__-__');
       });
 
+      it('should move cursor if backspace pressed on empty mask', () => {
+        cy.name('MINotControlledPhone')
+          .focusMasked()
+          .type('{rightArrow}{rightArrow}{rightArrow}{rightArrow}{rightArrow}{rightArrow}{rightArrow}')
+          .type('{backspace}{backspace}')
+          .type('981')
+          .should('have.value', '+7 (_98)-1__-__-__')
+          .clear()
+          .type('{rightArrow}{rightArrow}{rightArrow}{rightArrow}{rightArrow}{rightArrow}{rightArrow}')
+          .type('886')
+          .should('have.value', '+7 (___)-886-__-__')
+          .type('{rightArrow}{rightArrow}{rightArrow}{rightArrow}{rightArrow}{rightArrow}')
+          .type('{backspace}{backspace}')
+          .type('27')
+          .should('have.value', '+7 (___)-886-__-27')
+          .clear()
+          .should('have.value', '+7 (___)-___-__-__');
+      });
+
       it('should fill different masks', () => {
-        cy.get(`.${theme.wrapper} input`)
-          .eq(2)
+        cy.name('MIControlledPhone')
           .should('have.value', '+7 (800)-200-06-00')
           .focusMasked()
           .clear()
@@ -47,23 +62,20 @@ describe('MaskedInput', () => {
           .type('9818862798')
           .should('have.value', '+7 (981)-886-27-98')
           .closest('.demo-story')
-          .find('input')
-          .eq(1)
+          .name('MIControlledInsurance')
           .should('have.attr', 'placeholder', '___-___-___ __')
           .focusMasked()
           .clear()
           .type('12345678901')
           .should('have.value', '123-456-789 01')
           .closest('.demo-story')
-          .find('input')
-          .eq(3)
+          .name('MICarNumber')
           .should('have.attr', 'placeholder', 'Car number')
           .focusMasked()
           .type('AA12BB3456')
           .should('have.value', 'AA12BB3456')
           .closest('.demo-story')
-          .find('input')
-          .eq(4)
+          .name('MICreditCardNumber')
           .should('have.value', '6666-7777-8888-9999')
           .focusMasked()
           .clear()
@@ -72,8 +84,7 @@ describe('MaskedInput', () => {
       });
 
       it('should forbid non-mask chars', () => {
-        cy.get(`.${theme.wrapper} input`)
-          .eq(0)
+        cy.name('MINotControlledPhone')
           .focusMasked()
           .should('have.value', '+7 (___)-___-__-__')
           .type('ABC!@#$%^&*)_=+?/.<>,БЛА')
@@ -83,8 +94,7 @@ describe('MaskedInput', () => {
 
     describe('Validation', () => {
       it('should be invalid when isRequired and value is empty', () => {
-        cy.get(`.${theme.wrapper} input`)
-          .eq(0)
+        cy.name('MINotControlledPhone')
           .focusMasked()
           .blur()
           .closest(`.${theme.inputWrapper}`)
@@ -94,8 +104,7 @@ describe('MaskedInput', () => {
       });
 
       it('should be invalid when isRequired and value is not complete', () => {
-        cy.get(`.${theme.wrapper} input`)
-          .eq(0)
+        cy.name('MINotControlledPhone')
           .focusMasked()
           .type('1234')
           .blur()
@@ -106,8 +115,7 @@ describe('MaskedInput', () => {
       });
 
       it('should be valid when isRequired and value is complete', () => {
-        cy.get(`.${theme.wrapper} input`)
-          .eq(0)
+        cy.name('MINotControlledPhone')
           .focusMasked()
           .type('9818862798')
           .blur()
@@ -122,22 +130,19 @@ describe('MaskedInput', () => {
   describe('Rest', () => {
     describe('Controlled mode', () => {
       it('should clear and set value', () => {
-        cy.get(`.${theme.wrapper} input`)
-          .eq(2)
+        cy.name('MIControlledPhone')
           .should('have.value', '+7 (800)-200-06-00')
           .closest('.demo-story')
           .contains('Clear Value')
           .click()
           .closest('.demo-story')
-          .get(`.${theme.wrapper} input`)
-          .eq(2)
+          .name('MIControlledPhone')
           .should('not.have.value')
           .closest('.demo-story')
           .contains('Set Value')
           .click()
           .closest('.demo-story')
-          .get(`.${theme.wrapper} input`)
-          .eq(2)
+          .name('MIControlledPhone')
           .should('have.value', '+7 (981)-886-27-98');
       });
       it('should clear mask value', () => {