+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Updated .npmignore
 
 ### Fixed
+- MaskedInput: Fixed backspace
 - Fixed test coverage report for Windows